Improvements to the proof post-processor (#11821)
Includes the following changes:
(1) removes a misguided optimization where trusted proofs would be
collected during traversal. This leads to collecting trusted proofs that
are later removed by merging. This means that RARE reconstruction is
being run on (potentially many) trusted steps that will not be in the
final proof.
(2) Avoids adding trivial transitive steps in a case in
MACRO_SR_PRED_INTRO.
(3) Adds a stat.