Refactor distinct values program for datatypes and definition of DT_CONS_EQ_CLASH (#11897)
This is work towards eliminating the utility method `$get_fun` from the
CPC signature, which is not amenable to type checking.

This refactors the distinct values side condition for datatypes to avoid
`$get_fun`. It also makes `DT_CONS_EQ_CLASH` (which implements a very
similar side condition) reuse the distinct values side condition, which
is strictly more powerful.