Print statistics on timeout (#11922)
When trying to use the options `--stats-all` and `--stats-internal`,
they would not be printed when the timeout, specified via the `--tlimit`
option was reached. I am not sure whether there are other issues with
options being set in in the constructor of Env, but the `stats-all` and
`stats-internal` options have not been set when the StatisticsRegistry
is created:
https://github.com/cvc5/cvc5/blob/c167cfb2e57563a77158c56e89a22e05fbd02c34/src/smt/env.cpp#L65

When cvc5 hits a timeout, statistics are printed through
StatisticsRegistry::PrintSafe:
https://github.com/cvc5/cvc5/blob/c167cfb2e57563a77158c56e89a22e05fbd02c34/src/util/statistics_registry.cpp#L91

When cvc5 exits normally, a totally different path is taken for printing
statistics:
https://github.com/cvc5/cvc5/blob/c167cfb2e57563a77158c56e89a22e05fbd02c34/src/main/command_executor.cpp#L81

I don't know if this is the best way to handle it, but it does fix the
problem and was minimally invasive.

---------

Co-authored-by: Mathias Preiner <mathias.preiner@gmail.com>