Use model directly for arithmetic values in strings (#11891)
This simplifies string model construction by asking the model object
directly for models of length terms, instead of the Valuation class
which asks the appropriate theory solver.

The latter leads to spurious assertion failures when `--ee-mode=central`
is enabled since be may have chosen a term for representative that is
not shared.

Fixes https://github.com/cvc5/cvc5/issues/11890