Purification for nested sequence constants (#11865)
Constant sequences are represented as atomic terms. This can lead to
issues as the "subterms" of that sequence constant do not participate
e.g. in theory combination, nor are considered for model building.

This corrects this by a scheme which "purifies" nested sequence
constants recursively.

Fixes https://github.com/cvc5/cvc5-projects/issues/665