Remove deprecated table-based license key from pyproject.toml (#11952)
Table-based license keys of the form `license = {text =
"<license-string>"}` are deprecated in recent versions of Setuptools and
will no longer be supported after 2026-02-18. According to PEP 639, the
value for the license key must now be a simple string containing an SPDX
expression.

However, this new format is only supported by Setuptools >= 77.0.0,
which requires Python >= 3.9. To maintain compatibility with Python 3.7
and 3.8, as well as newer versions of Setuptools, this PR removes the
optional license key entirely and relies solely on the license files
included in the Python wheels to convey licensing information.

Refer to pypa/setuptools#4903 for a discussion of this change.