cadical: Do not notifySatLiteral on renotify_fixed. (#11998)
Fixes #11867.

When we get notified by CaDiCaL that a literal is fixed, it must already
be preregistered (literals that are not introduced on user level 0
cannot be fixed). Thus, we don't have to notifySatLiteral() when
renotifying the theory engine about fixed literals. This fix was tested
on whole SMT-LIB, incremental and non-incremental.