Do not filter output for the model-validation and unsat-cores tracks (#12012)
We decided to use the same call to cvc5 for the single-query,
model-validation, and unsat-cores tracks. PR #11882 implemented this by
using a copy of the same script for all three tracks. However, we only
want to filter the output for the single-query track. This PR updates
the scripts for the model-validation and unsat-cores tracks to account
for this difference.