Track has term incrementally for term db relevant (#12072)
This makes quantifiers roughly 10% faster on large industrial problems
in my initial tests.

In detail, we recently switched to `--term-db=relevant`, which requires
tracking which terms are relevant in the current SAT context. This makes
this computation maintained incrementally, instead of recomputed at each
full check.

This PR also includes a minor optimization to only update the relevant
domain when a ground term is non-congruent.

---------

Co-authored-by: Mathias Preiner <mathias.preiner@gmail.com>