Make NL more careful wrt theory combination (#11962)
This make the NL solver more careful with respect to theory combination.
In particular, certain parts of the NL solver (including the coverings
solver) are allowed to modify the model from the linear arithmetic
solver to satisfy nonlinear constraints. One rare issue that may arise
is if these changes identify 2 shared terms that are not equal before.
This makes it so that if 2 terms are identified in this way, then an
equality split is sent for those two terms.

Note that older versions (prior to
https://github.com/cvc5/cvc5/pull/7787) had a policy that was similar
but much less effective: shared terms would be split against their
value. That code was deleted in that PR (thereby possibly allowing
invalid models), this PR fixes the issue by introducing an improved
policy.

Fixes #10140. Fixes https://github.com/cvc5/cvc5/issues/12026.

Also adds a regression for a fixed issue. Fixes #9661.

Will test this on SMT-LIB.