Minor fixes to model construction for theory of datatypes (#12094)
This corrects the corner case of model construction for the theory of
datatypes.

In rare cases, an "instantiatied" constructor will be necessary to add
to the model, despite not being an asserted term.

For example, if `t = (cons (head t) (tail t))` where this fact did not
need to be shared, say, if `(head t)` is a term also belonging to theory
of datatypes.

This corrects the model construction in two ways: 
(1) `(head t)` and `(tail t)` need also be marked as relevant.  
(2) Constructor terms that are preregistered but not asserted should
*not* be marked as relevant.

The former ensures that doing congruence over selectors in the model
equality engine is not necessary.

This is in preparation for making our model construction more intuitive
(see https://github.com/cvc5/cvc5/pull/12093).