Do not do congruence over ordinary kinds in the model equality engine (#12093)
We should never rely on the core model object to do congruence over
specific theory operators. This keeps `HO_APPLY` as an exception to this
rule, which we need to congruence over based on how higher-order model
construction works.

It seems we have made theory-specific operators function kinds in the
model equality engine in the past to "patch" actual issues in model
construction.

This required further fixes to HO model construction to do
extensionality on functions that are infinite but have only finite
arguments, e.g. those (-> Bool Int).

This PR further includes benchmarks for DT that failed with this change
without https://github.com/cvc5/cvc5/pull/12094.