cmake: Update versioning scheme (#12116)
The current versioning scheme follows the format:
```
<version-in-last-tag>-dev.<number-of-commits-since-last-tag>.<commit-sha>[-modified]
```
This leads to two main practical issues:

1. To retrieve `<number-of-commits-since-last-tag>`, a clone of the
repository with access to the full Git history is required. This is not
always desirable, since it introduces unnecessary overhead. For example,
the GitHub `checkout` action retrieves only the latest commit by
default. When this information is not available, the build system does
not include other useful details such as the branch name or commit SHA
in the `CVC5_GIT_INFO` constant.

2. Even when the full history is available, the list of tags may not be.
This is common when a user or developer works on their own fork of the
repository.

This PR changes the format to:
```
<latest-release-version>-dev.<branch-name>@<commit-sha>[-modified]
```
This makes the version scheme more robust when the Git history or list
of tags is not available.

In addition, this PR preserves the old behavior of incrementing the
patch version by one when a development version is detected.