Optimizations for caching evaluation in RARE reconstruction (#12109)
This is a fixed version of https://github.com/cvc5/cvc5/pull/12005.

Makes two optimizations:
- NNF norm is cached, as it is expensive on large formulas and used as a
possible theory rewrite,
- Evaluations are not tried if the node does not rewrite to a constant.

The latter uncovered a bug in the arithmetic rewriter, namely that
`int.pow2` was considered expert (it is not expert, as it is used to
express BV rewrite rules), meaning it would not be rewritten. However,
it is part of `evaluate`, so pruning based on rewriting was too strict.

I will performance test this change.