java: Make CVC5ApiException a subclass of RuntimeException (#12141)
Currently, `CVC5ApiException` inherits from `Exception`, which is a
*checked exception* in Java. Checked exceptions must be declared in the
`throws` clause of any method that may throw them; failing to do so
prevents Java code using the method from catching the exception as a
`CVC5ApiException`. This PR changes `CVC5ApiException` to be a subclass
of `RuntimeException`, which is an *unchecked exception*. Unchecked
exceptions may optionally be listed in the *throws* clause, but omitting
them does not prevent code from catching them.

Fixes #12117