Add --inst-local option (#12121)
This option makes certain kinds of instantiation lemmas treated as
SAT-context dependent. We do so by:
1. making caches in the instantiate module SAT-context dependent,
2. instantiation lemmas are marked "local", which makes them SAT-context
dependent in the justification heuristic.
3. Local lemmas are guarded by a cache in the prop engine.

Note that we do not inform the SAT solver that the lemmas should be
considered SAT-context dependent currently, as the SAT solver doesn't
appear to be a bottleneck in most cases.

This option leads to significantly worse performance overall, but helps
in certain cases of interest. It is disabled by default.

The main downside is that instantiations are still only considered at
full effort, and thus are often added over and over again. Further
improvements will be to make instantiations considered more eagerly.