Treat nonlinear monomials as variables in the DIO solver (#12178)
The DIO solver was dropping equations containing non-linear polynomials
instead of handling their corresponding abstractions. This PR prevents
this behavior by treating non-linear monomials as variables. It also
simplifies the DIO solver by removing support for non-linear arithmetic,
which is no longer used.

This PR also disables the DIO solver for quantifier-free logics with
non-linear arithmetic, as it typically degrades performance on the
SMT-LIB benchmarks.

Further simplifications to the classes used by the DIO solver are
planned for future PRs.

---------

Co-authored-by: Andrew Reynolds <andrew.j.reynolds@gmail.com>