Add flatten monomials lemma schema (#12204)
This adds a new lemma schema to the NL solver and enables it by default.

This is +206-191 on unsat problems +883-848 on sat problems on SMT-LIB
logics with NL without FP, minimal performance difference on commonly
solved instances.

It also solves 9 benchmarks quickly on the verus set which previously
timed out.