Enable proof subtype elimination by default (#12198)
This enables `--proof-elim-subtypes` by default.

To do so, we make two changes:
1. Assertions given to the solver engine are processed by the subtype
elimination node converter. This means that if a user assertion used
mixed arithmetic, the proof will be in terms of a different preprocessed
form of that assertion. This change is necessary to ensure that proofs
do not contain mixed arithmetic (which cannot be avoided if the overall
assumptions have mixed arithmetic).

2. The subtype elimination node converter is relaxed so that real
division is overloaded (but not mixed). This change is motivated by the
fact that SMT-LIB rational constants are parsed as the real division of
integer constants and should not be considered as needing to be
processed by the subtype elimination node converter.

Also makes an optimization to the proof node converter to not update
proofs when it is unecessary to do so.