Basic Support for Logarithm with Base 2  (#12224)
This PR externalizes the INTS_LOG2 operator so that it can be used via
the API and textual interface.
For solving, we eliminate it in expand-definitions using int.pow2.

The cpc signature is updated to match the current semantics in cvc5,
namely: log2 of a negative number is 0.

Tests and traces are added, as well as minor fixes to existing comments.

---------

Co-authored-by: Yoni Zohar <yonizohar@Yonis-MacBook-Pro.local>