Eliminate shadowing at prerewrite (#12258)
Fixes #12252.

Since variable elimination is not sound for quantified formulas with
shadowing, this PR simplifies how variable shadowing is handled by
eagerly eliminating it at prerewrite.

Generalizes the proof reconstruction for lambda shadow elimination to
handle this case.

An assertion had to be deleted in this PR since now we may do shadow
elimination with `exists`.