[proofs] Fix cyclic proof issue in EqProof conversion
Sometimes the equality engine proof is silly in that an equality is derived
having itself as an assumption. While silly, this is not wrong, but when
building the proof, given the overwrite policy of assumptions, this would lead
to a cyclic proof.

A thorough way to avoid this issue is to not override assumptions, since in this
module assumptions are never supposed to be generated during proof conversion
other than original assumptions given to the module. So this commit makes the
construction of the proof more careful.

An alternative solution would be to guard the conversion in different points to
short-circuit cases in which you are rederiving something. But this is
error-prone and can be done in post-processing anyway.

Fixes #12240