Assign function values lazily in theory model (#12267)
This modifies the TheoryModel class so that models for functions are
computed on demand.

It also modifies the FMF module to compute function values when it is
invoked.

The motivation for this change is efficiency, as it function values are
not essentially in most cases during solving (apart from HO logics, and
when using mbqi or fmf), and moreover take significantly long time to
generate for large benchmarks. The performance overhead is particularly
in combination with lazy distinct handling (which requires models).
Using lazy function models reduces the overhead on a verification set of
interest from ~20% to ~4% with this PR.

This PR removes now obsolete methods that have been moved/updated in
TheoryModel.

---------

Co-authored-by: Haniel Barbosa <hanielbbarbosa@gmail.com>