Add lazy handling of distinct (#12136)
Currently we eagerly rewrite distinct to quadratically many
disequalities, which does not scale for large constraints.

This modifies the rewriter so that `distinct` with >5 children are not
rewritten.

This enables the distinct extension and updates the Eunoia signature to
treat `distinct` as `:arg-list` instead of `:pairwise`. It implements
the new rewrite rules for handling distinct and refactors the existing
rules to handle the new form of distinct.

It also handles the case where UF is not enabled and `distinct` is
provided, in which case it is eagerly eliminated at preprocessing. This
handles e.g. cases like where the logic is QF_BV and distinct should be
eagerly eliminated for the sake of eager bitblasting.

---------

Co-authored-by: Hans-Jörg <hansjoergschurr@users.noreply.github.com>