Update translation for ITE_EQ rule to Alethe (#12229)
Is the skolem section really the right place for this rule? It was there
already so I did not move it.

I did replace the original comment which I am not sure is a good idea
and I am happy to re-add. What other rules were you thinking of @HanielB
to be translated here? I'd need to give an example other than ITE.