Fix deployment of Maven artifacts (#12305)
The `pom.xml` file conditionally attaches platform-specific and
documentation artifacts only if they are generated during a build. This
is achieved by activating Maven profiles only when the corresponding
files are present in a directory determined by the `artifacts.dir` and
`javadoc.dir` properties. This is convenient for publishing only the
generated artifacts to a local Maven repository.

Unfortunately, when alternative values for these properties are
specified as command-line arguments, as we currently do in CI to publish
artifacts for all platforms, the new values are not taken into account
for Maven profile activation. This PR addresses the issue by reproducing
the expected directory configuration instead of providing alternative
paths.