Make solve-int-as-bv strict (#12324)
This preprocessing pass was only intended to work for quantifier-free
integer problems, this ensures we check for this.

Note when this pass is enabled, we are only able to answer "sat", in the
case of https://github.com/cvc5/cvc5/issues/12319, it was model unsound.

Also demotes this option to "expert".

Fixes https://github.com/cvc5/cvc5/issues/12319.